cv-model/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
bzip2=1.0.6=h14c3975_1002
ca-certificates=2018.11.29=ha4d7672_0
certifi=2018.11.29=py36_1000
cffi=1.11.5=py36h9745a5d_1001
cuda100=1.0=0
cycler=0.10.0=py_1
dbus=1.13.0=h4e0c4b3_1000
expat=2.2.5=hf484d3e_1002
ffmpeg=4.1=h6dce934_1000
fontconfig=2.13.1=h2176d3f_1000
freetype=2.9.1=h94bbf69_1005
gettext=0.19.8.1=h9745a5d_1001
glib=2.56.2=had28632_1001
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1001
gst-plugins-base=1.12.5=h3865690_1000
gstreamer=1.12.5=h0cc0488_1000
icu=58.2=hf484d3e_1000
intel-openmp=2019.1=144
jpeg=9c=h14c3975_1001
kiwisolver=1.0.1=py36h6bb024c_1002
libffi=3.2.1=hf484d3e_1005
libgcc-ng=7.3.0=hdf63c60_0
libgfortran-ng=7.2.0=hdf63c60_3
libiconv=1.15=h14c3975_1004
libopenblas=0.3.3=h5a2b251_3
libpng=1.6.36=h84994c4_1000
libprotobuf=3.6.1=hdbcaa40_1000
libstdcxx-ng=7.3.0=hdf63c60_0
libtiff=4.0.10=h648cc4a_1001
libuuid=2.32.1=h14c3975_1000
libxcb=1.13=h14c3975_1002
libxml2=2.9.8=h143f9aa_1005
matplotlib=3.0.2=py36h8a2030e_1001
matplotlib-base=3.0.2=py36h167e16e_1001
mkl=2019.1=144
mkl_fft=1.0.10=py36_0
mkl_random=1.0.2=py36_0
ncurses=6.1=hf484d3e_1002
nettle=3.4.1=h14c3975_1002
ninja=1.8.2=h6bb024c_1001
numpy=1.15.4=py36h7e9f1db_0
numpy-base=1.15.4=py36hde5b4d6_0
olefile=0.46=py_0
openblas=0.3.3=h9ac9557_1001
openh264=1.8.0=hdbcaa40_1000
openssl=1.0.2p=h14c3975_1002
pandas=0.23.4=py36h637b7d7_1000
pcre=8.41=hf484d3e_1003
pillow=5.4.1=py36h00a061d_1000
pip=18.1=py36_1000
protobuf=3.6.1=py36hf484d3e_1001
pthread-stubs=0.4=h14c3975_1001
pycparser=2.19=py_0
pyparsing=2.3.1=py_0
pyqt=5.6.0=py36h13b7fb3_1008
python=3.6.7=hd21baee_1001
python-dateutil=2.7.5=py_0
pytorch=1.0.0=py3.6_cuda10.0.130_cudnn7.4.1_1
pytz=2018.9=py_0
qt=5.6.2=hf70d934_9
readline=7.0=hf8c457e_1001
setuptools=40.6.3=py36_0
sip=4.18.1=py36hf484d3e_1000
six=1.12.0=py36_1000
sqlite=3.26.0=h67949de_1000
tk=8.6.9=h84994c4_1000
torchvision=0.2.1=py_2
tornado=5.1.1=py36h14c3975_1000
tqdm=4.29.1=py_0
wheel=0.32.3=py36_0
x264=1!152.20180717=h14c3975_1001
xorg-libxau=1.0.8=h14c3975_1006
xorg-libxdmcp=1.1.2=h14c3975_1007
xz=5.2.4=h14c3975_1001
zlib=1.2.11=h14c3975_1004
